requirements.txt
# Core Dependencies
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.23.0
matplotlib>=3.6.0

# Optional: For headless environments
# opencv-python-headless>=4.8.0
